split up NESTMLPrinter
