+++ pynestml/codegeneration/printers/cpp_variable_printer.py
@@ -34,9 +34,9 @@ class CppVariablePrinter(VariablePrinter):
         :param variable_name: a single name.
         :return: a string representation
         """
-        differential_order = variable_name.count("\"")
+        differential_order = variable_name.count("'")
         if differential_order > 0:
-            return variable_name.replace("\"", "").replace("$", "__DOLLAR") + "__" + "d" * differential_order
+            return variable_name.replace("'", "").replace("$", "__DOLLAR") + "__" + "d" * differential_order
 
         return variable_name.replace("$", "__DOLLAR")
