+++ pynestml/codegeneration/printers/nest_variable_printer.py
@@ -21,8 +21,6 @@
 
 from __future__ import annotations
 
-from pynestml.utils.ast_utils import ASTUtils
-
 from pynestml.codegeneration.nest_code_generator_utils import NESTCodeGeneratorUtils
 from pynestml.codegeneration.printers.cpp_variable_printer import CppVariablePrinter
 from pynestml.codegeneration.printers.expression_printer import ExpressionPrinter
